Avoiding warnings on integer conversions and handling overflows better

+++ test/common.h
@@ -30,7 +30,7 @@ auto readAll(const auto & readable) -> std::string {
         auto offset = res.size();
         constexpr size_t chunk = 5;
         res.resize(offset + chunk);
-        auto readCount = readFile(readable, res.data() + offset, chunk);
+        auto readCount = size_t(readFile(readable, res.data() + offset, chunk));
         res.resize(offset + readCount);
         if (readCount == 0)
             break;
